third_party/blink/web_tests/external/wpt/cookies/attributes/attributes-ctl.sub-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Cookie with %x9 in Domain attribute value is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9domain=t" but got ""
[FAIL] Cookie with %x9 in Path attribute value is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9path=t" but got ""
[FAIL] Cookie with %x9 in Max-Age attribute value is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9maxage=t" but got ""
[FAIL] Cookie with %x9 in Expires attribute value is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9expires=t" but got ""
[FAIL] Cookie with %x9 in Secure attribute is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9secure=t" but got ""
[FAIL] Cookie with %x9 after Secure attribute is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9secure2=t" but got ""
[FAIL] Cookie with %x9 in HttpOnly attribute is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9httponly=t" but got ""
[FAIL] Cookie with %x9 in SameSite attribute value is handled correctly.
  assert_equals: The cookie was set as expected. expected "test9samesite=t" but got ""
Harness: the test ran to completion.
